src/requirements.txt
PyYAML==6.0.1
chardet==5.2.0
customtkinter==5.2.2

presidio-analyzer==2.2.356
presidio-anonymizer==2.2.356

# ★ここが本丸：GiNZA(ja_ginza 5.2.0) と整合する spaCy を固定
spacy==3.7.5
ginza==5.2.0
ja-ginza==5.2.0

# PyInstaller/数値系の事故を避ける（Windowsで特に安定）
numpy==1.26.4

PyMuPDF==1.24.9
python-docx==1.1.2
lxml==5.2.2

pyinstaller==6.10.0
